fix: step

+++ tensor/internal/cudatensor/cuda_c/cudatensor.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 double *Full(size_t n, double value);
-double *Eye(size_t n, int d);
+double *Eye(size_t n, size_t d);
 double *RandU(size_t n, double l, double u);
 double *RandN(size_t n, double u, double s);
 double *Of(size_t n, const double input_data[]);
